renamed src/view/site-menu.js to src/view/menu.js
